// core/sj-frontend/src/client/app/services/service.ts
export class Service {
    type:string;
    name: string;
    description: string;
    keyspace: string;
    namespace: string;
    provider: string;
    'metadata-provider': string;
    'metadata-namespace': string;
    'data-provider': string;
    'data-namespace': string;
    'lock-provider': string;
    'lock-namespace': string;
    'zk-provider': string;
    'zk-namespace': string;
    index: string;
    login: string;
    password: string;
}